+++ app/src/main/java/com/u9porn/data/network/AppApiHelper.java
@@ -52,6 +52,7 @@ import com.u9porn.parser.ParsePxgav;
 import com.u9porn.parser.ParseV9PronVideo;
 import com.u9porn.parser.v9porn.VideoPlayUrlParser;
 import com.u9porn.rxjava.RetryWhenProcess;
+import com.u9porn.ui.porn9video.play.PlayVideoPresenter;
 import com.u9porn.utils.AddressHelper;
 import com.u9porn.utils.UserHelper;
 
@@ -63,6 +64,9 @@ import javax.inject.Inject;
 import javax.inject.Singleton;
 
 import io.reactivex.Observable;
+import io.reactivex.ObservableEmitter;
+import io.reactivex.ObservableOnSubscribe;
+import io.reactivex.functions.Function;
 import io.rx_cache2.DynamicKey;
 import io.rx_cache2.DynamicKeyGroup;
 import io.rx_cache2.EvictDynamicKey;
@@ -202,20 +206,30 @@ public class AppApiHelper implements ApiHelper {
     public Observable<VideoResult> loadPorn9VideoUrl(String viewKey) {
         String ip = addressHelper.getRandomIPAddress();
         //因为登录后不在返回用户uid，需要在此页面获取，所以当前页面不在缓存，确保用户登录后刷新当前页面可以获取到用户uid
-        return v9PornServiceApi.getVideoPlayPage(viewKey, ip, HeaderUtils.getIndexHeader(addressHelper))
-                .map(html -> videoPlayUrlParser.parseVideoPlayUrl(html, user));
-
-
-//        Map<String, String > headers = new HashMap<String, String>();
-//        headers.put("X-Forwarded-For",ip);
-//        headers.put("Referer", HeaderUtils.getIndexHeader(MyApplication.getInstance().getAddressHelper()));
-//        MyApplication.getInstance().getWebView().loadUrl(MyApplication.getInstance().getAddressHelper().getVideo9PornAddress()+"view_video.php"+"?viewkey="+viewKey);
-//        MyApplication.getInstance().getWebView().addJavascriptInterface(new InJavaScriptLocalObj(){
-//            @JavascriptInterface
-//            @Override
-//            public void showSource(String html) {
-//            }
-//        },"local_obj");
+//        return v9PornServiceApi.getVideoPlayPage(viewKey, ip, HeaderUtils.getIndexHeader(addressHelper))
+//                .map(html -> videoPlayUrlParser.parseVideoPlayUrl(html, user));
+        return Observable.create(new ObservableOnSubscribe(){
+            @Override
+            public void subscribe(ObservableEmitter emitter) throws Exception {
+                Map<String, String > headers = new HashMap<String, String>();
+                headers.put("X-Forwarded-For",ip);
+                headers.put("Referer", HeaderUtils.getIndexHeader(addressHelper));
+                MyApplication.getInstance().getWebView().addJavascriptInterface(new InJavaScriptLocalObj(){
+                    @JavascriptInterface
+                    @Override
+                    public void showSource(String html) {
+                        Logger.t(TAG).d("Webview 取回来的html source: "+html);
+                        emitter.onNext(html);
+                    }
+                },"local_obj");
+                MyApplication.getInstance().getWebView().loadUrl(MyApplication.getInstance().getAddressHelper().getVideo9PornAddress()+"view_video.php"+"?viewkey="+viewKey);
+            }
+        }).map(new Function<String,VideoResult>() {
+            @Override
+            public VideoResult apply(String html) throws Exception {
+                return videoPlayUrlParser.parseVideoPlayUrl(html, user);
+            }
+        });
     }
 
     @Override
